requirements.txt
colorama==0.3.9
decorator
django
djangorestframework
django-cors-headers
editdistance
fakeredis
gunicorn
psycopg2
py==1.4.34
pytest==3.1.2
redis
requests
sqlalchemy
testing.common.database==2.0.3
-e git+https://github.com/tk0miya/testing.postgresql.git@c81ded434d00ec8424de0f9e1f4063c778c6aaa8#egg=testing.postgresql
Unidecode==0.4.20